early implementation of save and restore state for the Movie List Activity

+++ app/src/main/java/pdm_1718i/yamda/data/server/DataMapper.kt
@@ -27,7 +27,7 @@ fun mapToMovie(jsonMovie : JSONObject) : Movie {
                 result.release_date,
                 result.id,
                 result.title,
-                result.vote_average,
+                result.vote_average.toFloat(),
                 result.budget,
                 result.genres.map {
                     Genre(it.name)
